+++ packages/core/src/app/address/AddressFormModal.tsx
@@ -95,11 +95,13 @@ const AddressFormModal: FunctionComponent<AddressFormModalProps> = ({
     onRequestClose,
     ...addressFormProps
 }) => {
-    const { themeV2 } = useThemeContext();
+    const { enhancedThemeV1 } = useThemeContext();
 
     return (
         <Modal
-            additionalModalClassName={classNames('modal--medium', 'modal--address', { themeV2 })}
+            additionalModalClassName={classNames('modal--medium', 'modal--address', {
+                enhancedThemeV1,
+            })}
             header={
                 <ModalHeader>
                     <TranslatedString id="address.add_address_heading" />