+++ packages/core/src/app/customer/EmailLoginForm.tsx
@@ -56,7 +56,7 @@ const EmailLoginForm: FunctionComponent<
     isFloatingLabelEnabled,
     values: { email: formEmail },
 }) => {
-    const { themeV2 } = useThemeContext();
+    const { enhancedThemeV1 } = useThemeContext();
 
     const modalHeaderStringId = useMemo(() => {
         if (emailHasBeenRequested) {
@@ -196,7 +196,7 @@ const EmailLoginForm: FunctionComponent<
     return (
         <Modal
             additionalBodyClassName="modal--withText"
-            additionalModalClassName={classNames('modal--medium', { themeV2 })}
+            additionalModalClassName={classNames('modal--medium', { enhancedThemeV1 })}
             header={
                 <ModalHeader>
                     <TranslatedString id={modalHeaderStringId} />