+++ packages/core/src/app/order/OrderSummary.tsx
@@ -118,7 +118,7 @@ const OrderSummary: FunctionComponent<OrderSummaryProps & OrderSummarySubtotalsP
                     storeCurrencyCode={storeCurrency.code}
                 />
                 {isTotalDiscountVisible && totalDiscount && (
-                    <div className="total-savings">
+                    <div className="total-savings optimizedCheckout-contentSecondary">
                         <TranslatedHtml
                             data={{ totalDiscount: currency.toCustomerCurrency(totalDiscount) }}
                             id="redeemable.total_savings_text"