+++ packages/core/src/app/shipping/MultiShippingFormFooter.tsx
@@ -34,7 +34,7 @@ const MultiShippingFormFooter: FunctionComponent<ShippingFormFooterProps> = ({
 
             <div className="form-actions">
                 <Button
-                    className="body-bold"
+                    className="optimizedCheckout-contentPrimary body-bold"
                     disabled={shouldDisableSubmit}
                     id="checkout-shipping-continue"
                     isLoading={isLoading}