+++ packages/core/src/app/shipping/SingleShippingForm.tsx
@@ -90,7 +90,7 @@ const SingleShippingForm: React.FC<
     const {
         shipping: { hideBillingSameAsShippingCheck },
     } = useCapabilities();
-    const { themeV2 } = useThemeContext();
+    const { enhancedThemeV1 } = useThemeContext();
     const {
         consignments,
         deinitializeShippingMethod: deinitialize,
@@ -289,7 +289,7 @@ const SingleShippingForm: React.FC<
 
     const shouldShowBillingSameAsShipping =
         !hideBillingSameAsShippingCheck &&
-        !themeV2 &&
+        !enhancedThemeV1 &&
         !PAYMENT_METHOD_VALID.some((method) => method === methodId);
 
     return (